TypeScript Cleanups and Fixes

+++ genbindings.py
@@ -288,7 +288,7 @@ def java_c_types(fn_arg, ret_arr_len):
             fn_arg = fn_arg[8:].strip()
         else:
             java_ty = consts.ptr_native_ty
-            c_ty = "int64_t"
+            c_ty = "intptr_t"
             arr_ty = "uintptr_t"
             rust_obj = "uintptr_t"
             fn_arg = fn_arg[9:].strip()
